Made ThreadList unnecessary in most applications: Thread has addition and multiplication operations, which make a ThreadListWrapper. Adjusted Workload constructors to take either a Thread or ThreadListWrapper.
Moved key/value generation and sizing methods and data into Operation.
That is needed because they may be dependent on the Table.
Added support for Table.key_size and Table.value_size to simplify applications.

Added THROW macro.

Fixed static counts of operations reported at the beginning of a run, they
were not taking into account any multiplier of operations.

C++ exceptions occurring in a thread were not properly caught.  Now the
'run' method catches these and rethrows the first one, so it can be converted
to a Python exception.