WT-3339 The CURSOR_UPDATE_API_CALL macro will dump core on a NULL btree handle
This is a problem for LSM or any non-standard data source that calls
CURSOR_UPDATE_API_CALL with an in-memory configuration and a NULL btree
handle.

Split CURSOR_UPDATE_API_CALL into two versions, one of which expects a
btree handle, one of which doesn't.

Rename the argument to TXN_API_CALL_NOCONF from "bt" to "dh", it's a
data handle in that macro, not a btree handle.