Switch to a spinlock whenever we need to lock a page.
Previously we had a read/write lock in the __wt_page structure that was
only ever acquired in write mode, plus a spinlock in the page->modify
structure.  Switch to using the spinlock for everything.

The only slight downside of this change is that we can no longer
precisely determine whether a page is locked based on the status of the
spinlock (since another page sharing the same lock could be holding it
in the places where we used to check).  Since that was only ever used
for diagnostic / debugging purposes, I think the benefit of the change
outweighs this issue.

Note also that this change fixes a bug where a failure during
`__wt_curfile_create` caused a data handle to be released twice.  This
is caught by the sanity checking assertions in the new read/write lock
code.