Allocate a spinlock for each modified page.
Using shared page locks for mulitple operations that need to lock a page
(including inserts and reconciliation) resulted in self-deadlock when
the lookaside table was used.  That's because reconciliation held a page
lock, then caused inserts to the lookaside table, which acquired the
page lock for a page in the lookaside table.  With a shared set of page
locks, they could both be the same lock.

Switch (back?) to allocating a spinlock per modified page.  Earlier in
this ticket we saved some space in __wt_page, so growing
__wt_page_modify is unlikely to be noticeable.