WT-3356 Use atomic reads of rwlocks.
Previously we had some conditions that checked several fields within a
rwlock by indirecting to the live structure.  Switch to always doing a
read of the full 64-bit value, then using local reads from the copy.

Otherwise, we're relying on the compiler and the memory model to order
the structure accesses in "code execution order".  That could explain
assertion failures and/or incorrect behavior with the new rwlock
implementation.