Cleanup and fixes.
Try to update the pinned timestamp if it is behind the the oldest
timestamp.  The check on timestamps being stable is slightly different
to the check on transactions being stable because we know transactions
are committed before we check timestamps.

Stop pretending we have two different rwlocks protecting transaction
state: there's really only one for now.