WT-3389 Restructure split code to hold a split generation for the entire operation.
Change insert, multi-page and reverse splits to acquire and hold a split
generation for the entire operation. This is intended to ensure internal
pages split by other threads cannot be evicted from underneath the current
split. This is intended to avoid problems like the one in WT-3373, where
an internal page was evicted during an attempt to split into it, leading
to an access violation.

These functions will already acquire and hold a split generation at some
point, so there's no performance cost with holding it for a longer period
(other than potentially failing to evict internal pages as soon as would
otherwise be possible). Additionally, we should be able to simplify some
of the underlying code.