__split_internal_lock() gets a hazard pointer on the child's parent page to avoid parent eviction. The problem that's meant to solve is that while the existence of the child page prevents parent page eviction initially, once the parent's page index is updated, the child may no longer be a child of that parent, the parent might have no children, and so could potentially be evicted.
I think that code is a leftover from older versions of the split code.
We currently hold a page lock on the parent page that prevents eviction
if the parent page is dirty, because eviction would require the page be
reconciled/cleaned. That should be safe: if the child page is moved so
the parent could be evicted, then the parent page must also have been
marked dirty.

If I'm wrong... there's an alternate version of this change we could
use.  Instead of getting a hazard pointer on the parent (which is a
relatively expensive operation that can fail, requiring the usual error
handling and cleanup), we should set the parent page's split generation
to be at least equal to our thread's split generation, because that
ensures the parent cannot be evicted: to evict, the child's WT_REF has
to move, and to move a WT_REF requires updating the parent page's split
generation.