Add a worker function for checking for illegal operations in in-memory configurations.
Don't allow backup cursors to be opened in in-memory configurations.
I originally started looking at this because backup cursors call the
schema-worker function, and we can get into trouble in that function
if we're in an in-memory configuration, because it closes open handles,
which attempts to checkpoint modified trees, and that's not possible
in in-memory configurations. However, backup cursors don't trigger
the problem because they don't specify a file-function argument to
schema-worker -- but, backup cursors don't make sense in in-memory
configurations, make them fail regardless.
It's actually safe to do so (we don't specify a file-function argument
to the schema-worker function
up in the schema-worker function