WT-3402 Move cached overflow records to the update list.
Cached overflow records are large value items that are being written
after being updated in a committed transaction (which implies freeing
and the potential reuse of their backing disk blocks), but which also
must be kept available because an older reader in the system may want
to read the previous value.

Historically, we maintained a separate cache of these overflow values
in the WT_PAGE_MODIFY.WT_OVFL_TXNC structure, with a set of functions
to maintain that list. The reason was because we wanted to be able to
remove overflow items from the cache based on updated transactional
information, that is, once any older readers exited the system, we could
discard the overflow values from the cache. Since that code was written,
we've added support for removing items from key/value WT_UPDATE lists
when they're no longer needed by earlier readers in the system, making
the WT_UPDATE lists sufficient to replace the separate overflow values
cache.

Rather than enter overflow values into the separate cache, append the
value to the WT_UPDATE list, with an impossibly low transaction ID to
ensure global visibility.