There's special-case code in reconciliation to handle an overflow value being removed, the overflow value being cached because of an older reader in the system, and then the page going through an update/restore reconciliation.
The idea is that if we cache an overflow value for an older reader in
the system, then do a page update/restore eviction, re-instantiating the
page in-memory means we'll no longer find the removed overflow value in
the cache because the index for the cache is the on-page disk address
of the overflow value, which doesn't appear in a re-instantiated page.

The fix was to take a copy of the original overflow value from the cache
and append it to the key's WT_UPDATE list.

With the move of the overflow cache from a separate list to the key's
WT_UPDATE list, that special-case code is no longer needed. Now, if an
overflow value is removed and cached, the cache is the key's WT_UPDATE
list, and subsequent reconciliations will always see a globally visible
entry.