When we moved the overflow values out of the page cache and into the update record, it broke variable-length column-store overflow values with RLE greater than 1.
Imagine an overflow value with an RLE greater than 1. We append a copy
to the end of an update chain, but it's possible it's the overflow value
for more than one record, and appending it to the end of one record's
update chain means a subsequent enter of a globally visible value to
one of the records would allow the truncation of the overflow chain that
leaves other records without a value.

The fix it that, if appending such an overflow record, set the transaction
ID to the first possible transaction ID. That ID is old enough to be
globally visible, but we can use it as a flag if an update record cannot
be discarded when truncating an update chain.

This fixes bugs already merged into the develop branch, specfically
ticket WT-3402 (PR #3493), commit 62fef4f should have had these changes.