Replace the uint8_t array with uint64_t values for 8-byte timestamps, ensuring we don't call memset/memcmp/memcpy unless it's necessary.
Convert any routine that operates on or returns a timestamp to take a
pointer to a timestamp in all cases.

Hide the magic in the WT_TIMESTAMP macro: in the case of an 8-byte
timestamp, it returns the address of the timestamp, in the case of an
array, it returns the declaration of the timestamp (also an address).
Add macros to call the underlying timestamp set and comparison functions,
which themselves call WT_TIMESTAMP.