WT-3380 Remove packed attribute from WT_UPDATE. (#3523)
* WT-3380 Remove packed attribute from WT_UPDATE.

Add 3B of data declaration at the end of the WT_UPDATE structure. That
way we don't have to pack the structure to avoid wasting data bytes,
and we don't have to use a macro to identify the start of the data.

* Locate the timestamp in the WT_UPDATE structure depending on its
alignment, to avoid padding.

* I lost a change, the size of the WT_UPDATE structure has to reflect the
size of the timetamp.

* Change the __wt_timestamp_t union into a structure so the compiler doesn't
insert padding in the middle of the WT_UPDATE structure (the existence of
the uint64_t in the union causes the whole thing to be aligned, even if we
never access it).

Incorporate Michael's change to replace sizeof(WT_UPDATE) with WT_UPDATESIZE,
the compiler is padding the structure at the end and we need to ignore that.

* If there's no union, we can't reach into it and get the field, take
the address like we do everywhere else.

* Simplify size calculations for WT_UPDATE.

In particular, go back to having WT_UPDATE_SIZE match the size of a
WT_UPDATE excluding the payload data.  That means the declared size of
the data array in WT_UPDATE is no longer special.

* Switch from a 3-byte array to a C99 flexible array member.

There is no longer anything special about the 3 byte array: since the
timestamp can be any size, there is no guarantee it makes a WT_UPDATE
any nicely rounded size.  Compilers enforce some rules around how
flexible array members can be used: we should consider switching our
other uses of structs with trailing arrays (in a separate ticket).

* Remove some more unnecessary casts now that upd->data is typed.