WT-3458 Checkpoint timestamps only apply to non-logged tables.
When logging is enabled for a table, that indicates that updates should
become durable immediately, since they will be unconditionally rolled
forward by recovery after a crash.  Since checkpoints with a timestamp
are intended to create a version on disk that does not include newer
changes, it doesn't make sense with immediately durable tables.