WT-3497 Improve logging message when hitting the WT session limits
Three places where we're using ENOMEM for generic "out of resource"
returns (transaction IDs, sessions, transaction notify entries), switch
to EAGAIN ("Resource temporarily unavailable") instead.

That's not exactly correct either (at least in the case of transaction
IDs, because we're never going to recover), but it's better than ENOMEM.

Add a verbose error message if transaction notify is called twice, it
should never happen, but we'll want to see an error message if it does.