Rework reconciliation to cache overflow records required by readers as part of selecting the WT_UPDATE to be written. That way we know the cached record is associated with the appropriate update.
Instead of caching a single copy of an on-page overflow value that's
being removed, and referencing it from both the WT_OVFL_TRACK structure
(for existing reader threads), and from all of the WT_UPDATE affected
chains, cache a complete copy of the on-disk overflow value itself
everywhere it might be needed. This requires 2x the memory when updating
row-store overflow tables, and potentially more for column-store tables
with RLEs greater than 1, but it has the virtue of working.
Additionally, the memory used should be discarded relatively quickly if
there's cache pressure: from both the update chains and from the
WT_OVFL_TRACK structure whenever we attempt to evict the page.