WT-3514 Fix checkpoint's use of stable_timestamp.
Prior to this change, there was a window in between checkpoint reading
the stable timestamp and using it to start a transaction.  In that
window, the application could update timestamps, making it invalid for
checkpoint's transaction to start.

Remove the undocumented "read_timestamp" config to
WT_SESSION::checkpoint, and instead copy the stable timestamp directly
into checkpoint's transaction while holding the global transaction
state locked.