MSVC Doesn't provide __func__, it has __FUNCTION__.
There was a fix in the utility code, move that fix into the WiredTiger
include files so it's generally available.

Switch WiredTiger diagnostics that use __FILE__ to use __func__, it
gives us a better indication of the error location if we don't have a
file version (and takes up less text space, if anyone cares).