Fix a race between create / drop and open_cursor.
During a create in particular, we first create the "table:" entry in
the metadata, then (for simple table) create a "colgroup:" entry.  The
latter gets the table handle before all column groups are complete: the
handle is open (WT_DHANDLE_OPEN) is set, but cg_complete = false.

There was nothing stopping an application thread with a cached table
handle acquiring the handle readlock and seeing it in this state.  To
avoid this, we now get the table handle exclusive during create and
hold it until the create completes.  Nothing else can race because
create is holding the schema and table locks exclusive.

While debugging, I found a few places where table handle were not being
released correctly.