WT-3437 Improvements to tuning of number of eviction workers
Updated the eviction workers tune algorithm to improve the eviction rate
and minimize the impact of re-tuning on the IOs. 

* Remove just one eviction worker forcefully every retune cycle to 
  avoid the situation of less numbers of eviction workers evicting

* Perform retune more frequently i.e. every 25 seconds to increase the 
  dynamism of the parallel eviction workers.

* Reduce the time period of collecting the datapoints to 60ms and
  increase the number of datapoints so that there is sufficient data 
  about eviction rate to settle on a stable number of workers. This also
  minimizes the time spent in retuning.

While at it, remove the restriction of thread[0] doing the retune. Any
of the thread in the eviction worker thread group could be an eviction 
server and can do the retune.