WT-3572 format should limit runs to 6 hours.
If the timer is not configured and the number of operations are
configured, set the timer to 6 hours, that is, allow plenty of
time for the operations to complete.

If neither the timer or the number of operations is configured,
set the timer for a relatively short run, 30 minutes.

Change the primary operations loop to force a quit if we're
still waiting for threads 15 minutes after the timer expires.