WT-3223 Progress messages for checkpoint operation.
While checkpointing a table (tree), log verbose message every 20 seconds
indicating the amount of work done in that duration. It was also observed that
the step of deleting the chekpoints, that are no longer required, is consuming
considerable amount of time while checkpointing a tree. Hence added a log
message at this step too.