Don't check if a ref is in the tree when searching during restore.
We used to assume that split was the only place that restores updates to
a page, and during a split the ref is not yet linked into a tree.  That
isn't the case for lookaside refs, which remain in the tree but need
updates restored.  Use a flag to make explicit what callers expect.