WT-3599 reconciliation calculates block matching checksums too frequently.
Calculating checksums is expensive and reconciliation does it more often
than necessary. Don't bother calculating checksums for: bulk-loads (LSM
does those a lot), the last-block in any reconciliation or eviction, where
there wasn't a previous block write to check it against or a previous
match has already failed.

Use the compressed image checksum for block matching when raw compression
configured, it's the same test and it's a smaller image to checksum.