Fix snapshot isolation failures in test/format.
Format generates modify updates that are not idempotent since they
add/remove bytes from the middle of a record.  That means
update/restore eviction has to discard updates from the on-page version
and older or we could apply a modification twice and get an invalid
value.  That in turn means we can't do update/restore eviction during
checkpoints.