WT-3640 Fix bytes-read and bytes-written statistics.
The bytes-read statistic includes allocated memory, but doesn't include
not overflow pages. Move the calculation from the page-allocation code
(which isn't part of reading overflow records), into __wt_bt_read(),
which is the low-level "read this address" function. Use the actual
bytes returned from the block-manager instead of the in-memory allocated
size from the page-allocation routine.

The bytes-written statistic isn't updated atomically, and is the page's
in-memory disk size, not the bytes written (for example, that ignores
compression). Update the value atomically and use the bytes actually
passed to the block-manager instead of the page's in-memory size.