WT-3619 compaction interactions with checkpoints and eviction.
Change compaction to not allow the compaction thread to be tapped for
eviction. Once a thread is tapped for eviction, it can spend a long time
doing nothing else and the compaction thread holds a lock that blocks
checkpoint.

Compaction blocks checkpoints for an entire file walk (and in the case
of many collections in a database, it might block on each collection).
Change compaction to acquire/discard WT_BTREE.flush_lock for each page
it reviews, ensuring checkpoint wins and can proceed. If compaction
can't acquire WT_BTREE.flush_lock, return EBUSY to the driver routine,
which waits for checkpoint to complete before starting the next
compaction pass.

Change compaction to complain and give up if there's eviction pressure.
In the face of eviction pressure, there's no point in doing compaction,
we're just making a bad problem worse.