WT-3652 Skip unnecessary lookaside reads / writes.
Also fix reads during checkpoint so the expected read generation is
set, and checkpoint can attempt to clean up after itself.

Add a "lookaside_score" measuring the proportion of unstable updates in
cache (those required for historic reads), use that to determine when
to use the lookaside table rather than waiting for the cache to become
stuck.