WT-3673 opening the lookaside table can race with the statistics server
The changes in c227e05 introduced a race between the statistics server
and opening the lookaside table. We open the statistics server before
creating the lookaside table and c227e05 changed to setting the
WT_CONN_LAS_OPEN flag before creating the connection's lookaside table
session and cursor, as part of making the code that creates internal
sessions own the problem of when to actually create lookaside table
cursors.

Rename WT_CONN_LAS_OPEN to be WT_CONN_LAS_CONFIGURED, and change the
statistics cursor to check if the connection's lookaside table session
and cursor fields are set, rather than checking a flag. The statistics
server is the only code that needs to check if the lookaside table has
been opened, and it's simpler than adding a new flag state.