Rename WT_CONN_LAS_CONFIGURED to WT_CONN_LOOKASIDE_CONFIGURED, matching other flags.
Quit bundling lookaside cursor creation with internal session creation,
that allows us to use WT_CONN_LOOKASIDE_CONFIGURED as the absolute
arbiter as to whether or not the current instance has a lookaside table
because it's not set until we're fully configured.

Add a write-barrier when setting WT_CONN_LOOKASIDE_CONFIGURED so we
can't race with the statistics server.

Clear that flag when shutting down the lookaside table handle (it
shouldn't matter and it could still race, but it's cleaner).