We create pages in-memory in several few cases: root or internal page splits, insert splits, root page reconciliations, and rebalance or salvage operations. Evicting those pages doesn't mean that eviction is making progress.
Instead of flagging if page eviction doesn't count as eviction progress,
go the other way, flag if page eviction should count as eviction progress.

Change so the only pages we count as eviction progress are pages we read
in while eviction was enabled, or pages that replace pages that were read
in while eviction was enabled.