Fix eviction in service of checkpoints.
Previously, we tracked whether eviction was successful, and if so,
continued the checkpoint from after the evicted page.  That could skip
over pages in some cases (presumably if eviction caused a split).

Instead, simplify the loop to make eviction advisory.  If eviction
succeeds, it should leave the reference in a state where checkpoint can
skip over it quickly.  If eviction fails, it may still have written the
reference and leave it clean, saving work for checkpoint.  Either way,
checkpoint visits every reference in the tree regardless of splits.