WT-3720 flags macros cast flags to unsigned values, hiding warnings.
The reason for flags.py was to auto-generate flags values, but also to
handle the configuration problems we had in Berkeley DB where different
APIs had flags that shared a name. WiredTiger went a different direction
with APIs and we don't have flags to APIs that share a name. Continue to
auto-generate flags values, but move the flags into the #include files
with their fields.