WT-3751 Allow splits when no data is visible.
In particular, when reconciling insert lists on row store leaf pages, we
also need to check for splits.  Now that we are handling this case
correctly, we are seeing splits with empty page images, which requires
further changes in reconciliation to track the last key associated with
a page (even if that key wasn't written to the page image).