WT-3768 Add an update type to version on-page values.
Previously, we stored full copies of on-disk values in the lookaside
table as well as in the disk image when we stored values that were not
globally visible.

This branch uses a different approach, and stores a new update type
(called WT_UPDATE_BIRTHMARK) whose purpose is to give the on-page value
a transaction time when it becomes visible.  Readers that see a
birthmark should read the on-disk value.  Readers that skip over a
birthmark because it is not visible to them should not see the on-page
value (equivalent to reading a tombstone at the end of the update list).