WT-3720 Stop casting in flags macros and change flag auto generation. (#3789)
Avoid casting the the flag macros - doing so can hide coding errors.

The reason for flags.py was to auto-generate flags values, but also to
handle the configuration problems we had in Berkeley DB where different
APIs had flags that shared a name. WiredTiger went a different direction
with APIs and we don't have flags to APIs that share a name. Continue to
auto-generate flags values, but move the flags into the #include files
with their fields.