WT-3767 Avoid lookaside reads when history is not required for queries.
Related to WT-3211 and WT-3776, which implemented the semantic that a
WT_CURSOR.remove that starts with a positioned cursor, returns with a
positioned cursor, but a WT_CURSOR.remove that doesn't start with a
positioned cursor, resets the cursor on return.

WT_REF_LIMBO complicates the problem because we now have a page state
where we must release our page pin inside of WT_CURSOR.remove, and redo
the search, because WT_CURSOR.remove may not be able to actually update
the pinned page.

Do the best we can: If we start with a positioned cursor, and we let
go of our pinned page, use a successful search to return to a
"positioned" state. If we start with a positioned cursor, let go of our
pinned page, and the search fails, leave the key set in the cursor so
the cursor appears positioned to the application.