WT-3768 Avoid storing on-page values in the lookaside table. (#3812)
Previously, we stored full copies of on-disk values in the lookaside table as well as in the disk image when we stored values that were not globally visible.

This branch uses a different approach, and stores a new update type (called WT_UPDATE_BIRTHMARK) whose purpose is to give the on-page value a transaction time when it becomes visible.  Readers that see a birthmark should read the on-disk value.  Readers that skip over a birthmark because it is not visible to them should not see the on-page value (equivalent to reading a tombstone at the end of the update list).

* If doing update/restore reconciliation and we skip a birthmark update, we have to retain the on-page value.

* Don't discard the chosen update until we've updated the maximum on-page timestamp value.

* In some cases (tombstones), a key from the original page may not appear on the update-eviction created page, we cannot assert an exact match in all cases.

* Row-store doesn't write zero-length data items so we can't use birthmarks for them.