WT-3769 Lookaside table should enable prefix compression
Configure prefix compression on the lookaside table. Testing this change
failed, with lookaside table cursors returning out-of-order keys. The
problem is if we use a pinned page for the cursor remove, we may not be
able to call __wt_key_return() to return a key, we require a search for
that call to be guaranteed to work.

First, merge a version of WiredTiger branch #3841 (reference WT-3767).

Then, rework the __wt_btcur_remove() code to ignore the key when the
remove was done based on an already positioned cursor.