WT-3818 __rec_txn_read() code order cleanup
Move all chosen update correctness assertions immediately after the code
to choose an update.

Put the maximum timestamp tracking tests together.

Move the test for not taking a chosen update because it's a birthmark,
or we skipped a birthmark, before the test of whether or not we've made
progress by taking an update.

Don't split the all_visible test/check into two parts.  If the update
chain isn't completely visible, set the leave-dirty flag before doing
any other work.