In commit 00e1b59, the test for if a pinned page is useful was broken: The test was originally:
	return (F_ISSET(cbt, WT_CBT_ACTIVE) &&
	    (!eviction_ok || cbt->ref->page->read_gen != WT_READGEN_OLDEST));

in other words, returning true if there was an active page and either the
caller specified eviction wasn't OK at this time or the read-generation
wasn't set to the special "oldest" value.

That got changed to:

	if (cbt->ref->page->read_gen != WT_READGEN_OLDEST)
		return (false);

which returns the pinned page isn't useful unless it's been scheduled
for forced eviction, which is wrong. This is turning off almost all
page-pinned behavior, there's not much chance a page has been flagged
for forced eviction.