WT-3779 Add support for string formats with WT_CURSOR::modify (#3851)
* Reconciliation generates cursors to perform modify operations, but
doesn't initialize the value-format field. Use the underlying tree
format instead. This means a bunch of the original commit to drive
the cursor down into the modify routines is no longer necessary,
but I'm leaving it alone, I slightly prefer using a cursor in the
low-level routines to using a WT_ITEM.

* The current session may not reference an underlying btree (when calling
through the "standard" cursor functions, as opposed to the "file" cursor
functions). Change reconciliation to initialize the value-format of the
cursor from the underlying tree.