WT-3792 LSM version 1 metadata incompatibility
Commit 3a2f5bf introduced a backward-incompatible change into the
LSM metadata by switching the order the id and generation values
appear in the chunks list. The reason was because custom data sources
need both the chunk ID and generation number to create the chunk
name. Switch the metadata write ordering back to avoid the
incompatibility, but don't create the chunk name until we have both
the chunk ID and the generation..