WT-3809 Fix a bug in lookaside related to birthmarks (#3845)
Birthmarks specify the visibility of an on-page value.  If a birthmark is visible, the on-page value should be returned.  If a birthmark is *not* visible, then we should check for an older value that is visible.  If no value is visible and we skipped a birthmark, we should not return the on-page value (i.e., the result should be `WT_NOTFOUND`).