WT-3818 __rec_txn_read() code order cleanup (#3850)
* WT-3818 __rec_txn_read() code order cleanup

Move all chosen update correctness assertions immediately after the code
to choose an update.

Put the maximum timestamp tracking tests together.

Move the test for not taking a chosen update because it's a birthmark,
or we skipped a birthmark, before the test of whether or not we've made
progress by taking an update.

Don't split the all_visible test/check into two parts.  If the update
chain isn't completely visible, set the leave-dirty flag before doing
any other work.

* We have to check for complete visibility before clearing (based on
birthmarks), the update we've chosen, otherwise we can't close the
file.

* Move birthmark handling before checks for saved updates.

Check for progress with the selected update before dealing with
birthmarks (reverts part of the change).

Fold birthmark handling into a check with a single effect rather than
repeating code.