Enhance operation tracking visualization to add a navigator chart. (#3892)
Added a navigator chart to the top of every HTML page showing an interval of execution across all threads (trace files). Using the navigator chart, we can easily navigate to another interval (by clicking on that interval) without having to go back to the main page. Unified the legend, so we have just one for all trace files, and it appears on top of the HTML file for that interval.