Change the sweep code to set the isolation level to read-uncommitted.
Save/restore the session isolation level instead of blindly setting it.
I don't think this fixes any bugs, but we get called from enough places
(including with the sweep session), that I want to future-proof the
code.

Set the session isolation level before beginning the local transaction,
it's not unreasonable for the isolation level to be considered when
beginning a transaction.