WT-3891 Take the sweep walk rwlock whenever removing content from lookaside (#3920)
* WT-3891 Take the sweep walk rwlock whenever removing content from lookaside

__wt_las_cursor_position "reads" the key/value pair, and so has to be
inside the r/w lock, otherwise we can race with the sweep remove code.

* The lookaside code shouldn't be freeing WT_MULTI structures, they're owned
by reconciliation.

* Change the sweep code to set the isolation level to read-uncommitted.

Save/restore the session isolation level instead of blindly setting it.
I don't think this fixes any bugs, but we get called from enough places
(including with the sweep session), that I want to future-proof the
code.

Set the session isolation level before beginning the local transaction,
it's not unreasonable for the isolation level to be considered when
beginning a transaction.

* Re-factor __wt_las_remove_block into two parts, an external version that
allocates a cursor, and an internal version that is passed a cursor. It
makes error handling a lot cleaner, and I'm going to need that split when
I change the record counter decrement to be part of transaction commit.

* Wrap sweep's removal of lookaside table entries in a transaction.

* Acquire the write lock in __wt_las_sweep() before positioning the cursor,
otherwise we could race with the remove-block code.

Check specifically for WT_NOTFOUND errors after the calls from which we
expect them, instead of testing generally at the end of the function.

* If remove is running concurrently with insert, it's possible for records
to be removed before the insert transaction commit (remove is configured
read-uncommitted). Make sure increments stay ahead of decrements.

* Don't hard-code the buffer size, there's no reason to snprintf() a string
into a buffer.

* Don't increment/decrement the lookaside table counters unless the commit
succeeds.

* lint

* I didn't bother checking if the lookaside table decrements were non-zero,
it's a cheap check, do it in the inlined code (and the compiler should
compress out any duplicate check in any callers).

* Lint

* My commit "Check specifically for WT_NOTFOUND errors after the calls
from which we expect them, instead of testing generally at the end of
the function." was wrong, put it back the way it was.

* The other half of the change to move freeing the update list from
the lookaside table insert code into the reconciliation code.

* Review comments: make the two push/pop isolation function names mirror
each other.

* Review comments, clarify a comment.

* Review comments, fix a comment.

* More review comments, fix a couple of comments, clean up some whitespace.